+++ src/main/java/com/bitmart/api/request/spot/prv/v4/V4QueryOrderByClientOrderIdRequest.java
@@ -34,7 +34,6 @@ public class V4QueryOrderByClientOrderIdRequest extends CloudRequest {
     private Long recvWindow;
 
     /**
-     * url: POST https://api-cloud.bitmart.com/spot/v4/query/client-order
      * Query a single order based on the client order id
      */
     public V4QueryOrderByClientOrderIdRequest() {