+++ src/main/java/com/bitmart/api/request/spot/pub/SymbolsRequest.java
@@ -17,7 +17,6 @@ import lombok.experimental.Accessors;
 public final class SymbolsRequest extends CloudRequest {
 
     /**
-     * url: GET https://api-cloud.bitmart.com/spot/v1/symbols
      * Get a list of all trading pairs on the platform
      */
     public SymbolsRequest() {